% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_fact_sheet_html.R
\name{construct_fact_sheet_html}
\alias{construct_fact_sheet_html}
\title{construct_fact_sheet_html}
\usage{
construct_fact_sheet_html()
}
\description{
Adds the necessary templates for making PDF fact sheets at the Urban
Institute.Adds a .gitignore, pdf_fact_sheet template, LaTeX preamble, and
iteration script.
}
