% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_fact_sheet_pdf.R
\name{construct_fact_sheet_pdf}
\alias{construct_fact_sheet_pdf}
\title{construct_pdf_fact_sheet}
\usage{
construct_fact_sheet_pdf()
}
\description{
Adds the necessary templates for making PDF fact sheets at the Urban
Institute.Adds a .gitignore, pdf_fact_sheet template, and LaTeX preamble.
}
