% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_pdf_fact_sheet.R
\name{construct_pdf_fact_sheet}
\alias{construct_pdf_fact_sheet}
\title{construct_pdf_fact_sheet}
\usage{
construct_pdf_fact_sheet()
}
\description{
Adds the necessary templates for making PDF fact sheets at the Urban
Institute.Adds a .gitignore, pdf_fact_sheet template, and LaTeX preamble.
}
