% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_content.R
\name{use_content}
\alias{use_content}
\title{use_content}
\usage{
use_content(content)
}
\arguments{
\item{content}{A character string for the name of the content. The
only options are "header image" and "urban logo".}
}
\description{
Adds content to the specified directory.
}
