% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fact_sheet_pdf.R
\name{use_fact_sheet_pdf}
\alias{use_fact_sheet_pdf}
\title{use_fact_sheet_pdf}
\usage{
use_fact_sheet_pdf(name = "fact_sheet_pdf.Rmd", directory = NULL)
}
\arguments{
\item{name}{A character string for the name of the created template. Be sure
to end the name with ".Rmd".}

\item{directory}{A character string for the directory where the template
should be saved.}
}
\description{
use_fact_sheet_pdf
}
