% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_git_ignore_urbn.R
\name{use_git_ignore_urbn}
\alias{use_git_ignore_urbn}
\title{use_git_ignore_urbn}
\usage{
use_git_ignore_urbn(gitignore = NULL, open = FALSE)
}
\arguments{
\item{gitignore}{Selection of gitignore Current options are \code{"shiny"},
and \code{"fact_sheet_pdf"}.}

\item{open}{Open the newly created file for editing?}
}
\description{
\code{use_git_ignore_urbn} adds a custom .gitignore to the top directory.
}
