% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_iterate.R
\name{use_iterate}
\alias{use_iterate}
\title{use_iterate}
\usage{
use_iterate(directory = NULL)
}
\arguments{
\item{directory}{A character string for the directory where the template
should be saved.}
}
\description{
Adds a template called iterate.R that is used for iterating fact cheets
across tibbles.
}
