% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pdf_fact_sheet.R
\name{use_pdf_fact_sheet}
\alias{use_pdf_fact_sheet}
\title{use_pdf_fact_sheet}
\usage{
use_pdf_fact_sheet(name = "pdf_fact_sheet.Rmd", directory = NULL)
}
\arguments{
\item{name}{A character string for the name of the created template. Be sure
to end the name with ".Rmd".}

\item{directory}{A character string for the directory where the template
should be saved.}
}
\description{
use_pdf_fact_sheet
}
