% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_readme_urbn.R
\name{use_readme_urbn}
\alias{use_readme_urbn}
\title{use_readme_urbn}
\usage{
use_readme_urbn(name = "README")
}
\arguments{
\item{name}{Name of analysis to put at top of README}
}
\description{
Adds an Urban Institute specific README.
}
