% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_shiny_app.R
\name{use_shiny_app}
\alias{use_shiny_app}
\title{use_shiny_app}
\usage{
use_shiny_app()
}
\description{
\code{use_shiny_app} adds a template called app.R for creating R Shiny
applications. There is no name argument because the Urban Institute Shiny
server expects app.R.
}
