% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_web_report.R
\name{use_web_report}
\alias{use_web_report}
\title{use_web_report}
\usage{
use_web_report(name = "web_report.Rmd", directory = NULL)
}
\arguments{
\item{name}{A character string for the name of the created template.}

\item{directory}{A character string for the directory where the template
should be saved.}
}
\description{
Adds a web report template to the specified directory.
}
