% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_fact_sheet_html.R
\name{construct_fact_sheet_html}
\alias{construct_fact_sheet_html}
\title{construct_fact_sheet_html}
\usage{
construct_fact_sheet_html()
}
\description{
Adds the necessary templates for making html fact sheets at the Urban
Institute. Adds a .gitignore, html fact sheet template, CSS, iteration
script, and instructions.
}
