% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_fact_sheet_html.R
\name{construct_fact_sheet_html}
\alias{construct_fact_sheet_html}
\title{construct_fact_sheet_html}
\usage{
construct_fact_sheet_html(name = "fact_sheet_html.Rmd",
  directory = NULL, instructions = TRUE)
}
\arguments{
\item{name}{A character string for the name of the created template. Be sure
to end the name with ".Rmd".}

\item{directory}{A character string for the directory where the template
should be saved.}

\item{instructions}{Should a markdown of helpful instructions be included?
Defaults to \code{TRUE}.}
}
\description{
Adds the necessary templates for making html fact sheets at the Urban
Institute. Adds a .gitignore, html fact sheet template, CSS, iteration
script, and instructions.
}
