% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_slide_show.R
\name{construct_slide_show}
\alias{construct_slide_show}
\title{construct_slide_show}
\usage{
construct_slide_show()
}
\description{
Adds the necessary templates for making slide shows at the Urban
Institute.Adds a .gitignore, revealjs template, CSS styles, and
a header image.
}
