% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_web_report.R
\name{construct_web_report}
\alias{construct_web_report}
\title{construct_web_report}
\usage{
construct_web_report()
}
\description{
Adds the necessary templates for making web reports at the Urban
Institute.Adds a .gitignore, pdf_fact_sheet template, LaTeX preamble, and
iteration script.
}
