% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_css.R
\name{use_css}
\alias{use_css}
\title{use_css}
\usage{
use_css(stylesheet, directory = "www", open = FALSE)
}
\arguments{
\item{stylesheet}{Selection of stylesheet. Current options are \code{"shiny"}.}

\item{directory}{The directory where the template should be saved and opened.
The default "www" is preferable for most projects. Use NULL to save the
template to the project directory or working directory.}

\item{open}{Open the newly created file for editing?}
}
\description{
Adds a Cascading Style Sheet (CSS) to the specified directory
}
