% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fact_sheet_html.R, R/use_fact_sheet_pdf.R
\name{use_fact_sheet_pdf}
\alias{use_fact_sheet_pdf}
\alias{use_fact_sheet_pdf}
\title{use_fact_sheet_html}
\usage{
use_fact_sheet_pdf(name = "fact_sheet_pdf.Rmd", directory = NULL)

use_fact_sheet_pdf(name = "fact_sheet_pdf.Rmd", directory = NULL)
}
\arguments{
\item{name}{A character string for the name of the created template. Be sure
to end the name with ".Rmd".}

\item{directory}{A character string for the directory where the template
should be saved.}

\item{name}{A character string for the name of the created template. Be sure
to end the name with ".Rmd".}

\item{directory}{A character string for the directory where the template
should be saved.}
}
\description{
Adds an R Markdown template for creating html fact sheets at the Urban
Institute. For more information,
\href{https://medium.com/@urban_institute/iterated-fact-sheets-with-r-markdown-d685eb4eafce}{read here}.

Adds an R Markdown template for creating PDF fact sheets at the Urban
Institute. For more information,
\href{https://medium.com/@urban_institute/iterated-pdfs-with-r-markdown-144e2a6d6a1a}{read here}.
}
