% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_instructions.R
\name{use_instructions}
\alias{use_instructions}
\title{use_instructions}
\usage{
use_instructions(instructions)
}
\arguments{
\item{instructions}{A character string for the name of the created template.
Current options are \code{"fact_sheet_pdf"}, \code{"fact_sheet_html"}, \code{"revealks"},
\code{"shiny"}, and \code{"web_report"}.}
}
\description{
Adds instructions for constructed templates.
}
