% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_revealjs.R
\name{use_revealjs}
\alias{use_revealjs}
\title{use_revealjs}
\usage{
use_revealjs(name = "revealjs.Rmd")
}
\arguments{
\item{name}{A character string for the name of the created template. Use / to
specify a subdirectory like \code{name = "docs/revealjs.Rmd"}.}
}
\description{
use_revealjs
}
